+++ src/main/java/Homura/Deadline.java
@@ -1,14 +1,24 @@
+// How to use and create packages inspired by
+// https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
+// and
+// https://www.w3schools.com/java/java_packages.asp
+package Homura;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
-public class Event extends Todo {
+public class Deadline extends Todo {
     // Attributes + Getters and Setters ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     // Todo.str  description
     // Todo.bool isDone
-    private LocalDate sta;
-    private LocalDate end;
+    private LocalDate by;
 
+    // Inspired by ChatGPT using queries "Hi! Can I have some
+    // examples on how I can use LocalDate.Format() to format
+    // my dates in Java?" and "Hi! Can I have some samples of
+    // how LocalDate in Java can parse a string and return a
+    // LocalDate object?"
     public static final DateTimeFormatter dtfParse
             = DateTimeFormatter.ofPattern("yyyy-MM-dd");
     public static final DateTimeFormatter dtfToString
@@ -16,21 +26,18 @@ public class Event extends Todo {
     public static final DateTimeFormatter dtfToStorage
             = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
-    public LocalDate getSta() { return sta; }
-    public LocalDate getEnd() { return end; }
+    public LocalDate getBy() { return by; }
 
 
 
     // Constructors and Factory Methods ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-    public Event(String description, LocalDate sta, LocalDate end) {
+    public Deadline(String description, LocalDate by) {
         super(description);
-        this.sta = sta;
-        this.end = end;
+        this.by = by;
     }
-    public Event(String description, String staStr, String endStr) {
+    public Deadline(String description, String byStr) {
         super(description);
-        this.sta = LocalDate.parse(staStr, dtfParse);
-        this.end = LocalDate.parse(endStr, dtfParse);
+        by = LocalDate.parse(byStr, dtfParse);
     }
 
 
@@ -38,20 +45,19 @@ public class Event extends Todo {
     // String Methods ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     @Override
     public String toString() {
-        String ans = "[E]";
+        String ans = "[D]";
         if (getIsDone()) {
             ans += "[✓] ";
         } else {
             ans += "[ ] ";
         }
-        ans += getDescription()
-                + " (from: " + sta.format(dtfToString)
-                + " to: " + end.format(dtfToString) + ")";
+        ans += getDescription() + " (by: "
+                + by.format(dtfToString) + ")";
         return ans;
     }
-    public static Event fromStorageStr(String s) {
+    public static Deadline fromStorageStr(String s) {
         ArrayList<String> ss = HomuraUtils.split(s, Storage.DIVIDER);
-        Event ans = new Event(ss.get(2),ss.get(3),ss.get(4));
+        Deadline ans = new Deadline(ss.get(2),ss.get(3));
         if (ss.get(1).equals("1")) {
             ans.setIsDone(true);
         }
@@ -59,16 +65,15 @@ public class Event extends Todo {
     }
     @Override
     public String toStorageStr() {
-        // e | 0 or 1 | descr | sta | end
-        String ans = "e" + Storage.DIVIDER;
+        // d | 0 or 1 | descr | by
+        String ans = "d" + Storage.DIVIDER;
         if (getIsDone()) {
             ans += 1;
         } else {
             ans += 0;
         }
         ans += Storage.DIVIDER + getDescription()
-                + Storage.DIVIDER + sta.format(dtfToStorage)
-                + Storage.DIVIDER + end.format(dtfToStorage);
+                + Storage.DIVIDER + by.format(dtfToStorage);
         return ans;
     }
 }
@@ -79,21 +84,19 @@ public class Event extends Todo {
 
 // Recycling Bin ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 /*
-    public static Event parseUserInp(String inp) {
+    public static Deadline parseUserInp(String inp) {
         inp = inp.strip();
-        if (inp.length() <= 6) {
+        if (inp.length() <= 9) {
             throw new EmptyInputHomuraException("Todo", inp);
         }
         try {
-            inp = inp.substring(6);   // Remove the "event " in front
-            String[] splitInps = inp.split(" /from ");
+            inp = inp.substring(9);   // Remove the "deadline " in front
+            String[] splitInps = inp.split(" /by ");
             String descr = splitInps[0];
-            splitInps = splitInps[1].split(" /to ");
-            String sta = splitInps[0];
-            String end = splitInps[1];
-            return new Event(descr, sta, end);
+            String byStr = splitInps[1];
+            return new Deadline(descr,byStr);
         } catch (Exception e) {
-            throw new InvalidInputHomuraException("event", inp);
+            throw new InvalidInputHomuraException("deadline", inp);
         }
     }
  */
@@ -101,26 +104,35 @@ public class Event extends Todo {
 
 
     /**
-     * Converts a storage string representation to an Event.
+     * Converts a storage string representation to a Deadline.
      *
      * @param s The storage string.
-     * @return The Event.
+     * @return The Deadline.
      */
 
     /**
      * Converts to a String representation for storage.
-     * The current format is `e | 0 or 1 | descr | sta | end`.
+     * The current format is `d | 0 or 1 | descr | by`.
      *
      * @return The storage representation.
      */
 
 
 
+// TODO
+    /*
+    - Implement a fromStorageStr() function
+    - Replicate similar updates to Event.java
+    - Also implement fromStorageStr() in Todo.java
+    - Update how Storage.readTodos() words (likely deprecating
+      Storage.fromStr() in the process)
+     */
+
 /*
     @Override
     public String mark_str() {
         return indent + divider + '\n'
-            + indent + " Event marked as done" + '\n'
+            + indent + " Deadline marked as done" + '\n'
             + indent + " " + this.toString() + '\n'
             + indent + divider + '\n';
     }
@@ -129,7 +141,7 @@ public class Event extends Todo {
     @Override
     public String unmark_str() {
         return indent + divider + '\n'
-            + indent + "Event marked as not done" + '\n'
+            + indent + "Deadline marked as not done" + '\n'
             + indent + " " + this.toString() + '\n'
             + indent + divider + '\n';
     }
