+++ src/main/java/Homura/Playground.java
@@ -1,6 +1,13 @@
 // For testing code without using I Run Code Discord Bot
 // esp cause I can't fit all my classes in 2k chars
 
+// How to use and create packages inspired by
+// https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
+// and
+// https://www.w3schools.com/java/java_packages.asp
+package Homura;
+
+
 import java.util.*;
 
 public class Playground {
