+++ src/main/java/Homura/Storage.java
@@ -1,3 +1,9 @@
+// How to use and create packages inspired by
+// https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
+// and
+// https://www.w3schools.com/java/java_packages.asp
+package Homura;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
